refactor (1) Moved validate logic from Controller to Manager. (2) Remove some tests which are private methods.
