+++ src/main/java/racing/controller/RacingController.java
@@ -17,27 +17,13 @@ public class RacingController {
     }     
     
     public RacingManager makeNewGame(GameMakingInfo gameMakingInfo) {
-        makeSureUserInputs(gameMakingInfo);
         racingManager = new RacingManager(gameMakingInfo);
         return racingManager;
-    }
-            
-    private void makeSureUserInputs(final GameMakingInfo gameMakingInfo) {
-        if (!RacingValidator.isValidCarNames(gameMakingInfo.getCarNames())) {
-            throwException(ErrorMessage.INCORRECT_CAR_NAMES.getMessage());
-        }
-        if (!RacingValidator.isInTimeRange(gameMakingInfo.getTime())) {
-            throwException(ErrorMessage.INCORRECT_TIME.getMessage());
-        }
     }
     
     public void processGame() {
         racingManager.startGame();
         resultView.printMovingStatus(racingManager.getMovingHistory());
         resultView.printWinners(racingManager.getWinnerNames());
-    }
-    
-    private static void throwException(String errorMessage) {
-        throw new IllegalArgumentException(errorMessage);
     }
 }
