+++ src/main/java/racing/domain/RacingManager.java
@@ -2,7 +2,9 @@ package racing.domain;
 
 import static racing.common.RacingSettings.*;
 
+import racing.common.ErrorMessage;
 import racing.common.RacingSettings;
+import racing.common.RacingValidator;
 import racing.common.RandomNumberGenerator;
 import racing.vo.GameMakingInfo;
 
@@ -21,12 +23,22 @@ public class RacingManager {
     private int time;
     
     public RacingManager(GameMakingInfo gameMakingInfo) {
+        makeSureUserInputs(gameMakingInfo);
         this.cars = new Cars(Arrays.stream(gameMakingInfo.getCarNames().split(RacingSettings.CAR_NAME_SEPARATOR.getStr()))
             .map(Car::new)
             .collect(Collectors.toList())
         );
         time = gameMakingInfo.getTime();
         movingHistory = new ArrayList<>();
+    }
+    
+    private void makeSureUserInputs(GameMakingInfo gameMakingInfo) {
+        if (!RacingValidator.isValidCarNames(gameMakingInfo.getCarNames())) {
+            throwException(ErrorMessage.INCORRECT_CAR_NAMES.getMessage());
+        }
+        if (!RacingValidator.isInTimeRange(gameMakingInfo.getTime())) {
+            throwException(ErrorMessage.INCORRECT_TIME.getMessage());
+        }
     }
     
     public String getWinnerNames() {
@@ -46,5 +58,9 @@ public class RacingManager {
     
     public void setCars(Cars cars) {
         this.cars = cars;
+    }
+    
+    private static void throwException(String errorMessage) {
+        throw new IllegalArgumentException(errorMessage);
     }
 }
